kernel: ksud, core_hook: migrate ksud execution to security_bprm_check
This migrates ksud execution decision-making to bprm_check_security.
This requires passing proper argv and envp to a modified _ksud handler
aptly named 'ksu_handle_bprm_ksud'.

Introduces:
int ksu_handle_bprm_ksud(const char *filename, const char *argv1,
			const char *envp, size_t envp_len)

which is adapted from:
int ksu_handle_execveat_ksud(int *fd, struct filename **filename_ptr,
			     struct user_arg_ptr *argv,
			     struct user_arg_ptr *envp,
			     int *flags)

ksu_handle_bprm_ksud handles all the decision making, it decides when it is
time to apply_kernelsu_rules depending if it sees "second_stage".

For LSM hook, turns out we can pull out argv and envp from mm_struct.
The code in here explains itself on how to do it.

whole blob exists on arg_start to arg_end, so we just pull it out and grab next
array after the first null terminator.

as for envp, we pass the pointer and hunt for it on ksu_handle_bprm_ksud

as for hardening:
- proper checks (filtering, oob checks)
- GFP_ATOMIC kmalloc to avoid or lower blocking time
- copy_from_user_nofault + a copy_from_user fallback

My reasoning on adding a fallback is that on some devices a fault happens,
and it has copied garbled data. On my creation of this, I actually had to
lock that _nofault copy on a spinlock as a way to mimic preempt_disable/enable
without actually doing it. As per user reports, no failed _nofault copies anyway
but we have-to-have a fallback for resilience.

References:
- old version https://github.com/backslashxx/KernelSU/commit/6efcd8193e62d13a4e62cda0ce54d6770260c680
- bad usercopy https://github.com/backslashxx/KernelSU/issues/21

This now provides a small helper function, ksu_copy_from_user_retry, which explains
itself. First we attempt a _nofault copy, if that fails, we try plain.

With that, I also need to provide  an inlined copy_from_user_nofault for < 5.8.
The old probe_user_write for < 5.8 is NOT really _nofault as it does NOT have that
pagefault_disable + enable which is the hallmark of these _nofault functions.

While using strncpy_from_user_nofault was considered, this wont do, this will
only copy up to the first \0.

devlog:
https://github.com/ximi-libra-test/android_kernel_xiaomi_libra/compare/16e5dce9e7e...16c1f5f521a
https://github.com/backslashxx/mojito_krenol/compare/28642e60d7b...728de0c571c

References:
https://elixir.bootlin.com/linux/v4.14.1/source/include/linux/mm_types.h#L429
https://elixir.bootlin.com/linux/v4.14.1/source/include/linux/lsm_hooks.h
--

With this change, theres no need to use execve_ksud and execveat_ksud
handlers anymore.

On next commit, we clean or lock those up.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>