kernel: allowlist: escape persistent_allow_list to kthread
seems this has to be escaped to root, or else
[   84.532520] KernelSU: save_allow_list create file failed: -13

I'm not sure if task work fixes it for upstream, but we can't rely on it due to
this repo supporting ultra-legacy.

For this, we have to go with primitives. While this is might be a bit heavier,
a full kthread has its own context and we can escape it to what we want.

devlog:
	https://github.com/ximi-mojito-test/mojito_krenol/compare/69b47f1...21790e3
	https://github.com/backslashxx/KernelSU/compare/b2fd62d...54da9f2
	https://github.com/backslashxx/KernelSU/compare/81bff36...5ba798c

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>